net: lib: nrf_cloud: cleanup duplicate functions
Move duplicate functions to shared codec file.
Added brief comments to the codec header file.
Ref: CIA-382
