+++ subsys/net/lib/nrf_cloud/include/nrf_cloud_codec.h
@@ -105,32 +105,66 @@ int nrf_cloud_device_status_encode(const struct nrf_cloud_device_status * const 
 
 /** @brief Free memory allocated by @ref nrf_cloud_device_status_encode. */
 void nrf_cloud_device_status_free(struct nrf_cloud_data *status);
+
+/** @brief Frees memory allocated by @ref nrf_cloud_rest_fota_execution_parse */
 void nrf_cloud_fota_job_free(struct nrf_cloud_fota_job_info *const job);
 
+/** @brief Parses the response from a FOTA execution request REST call.
+ * If successful, memory will be allocated for the data in @ref nrf_cloud_fota_job_info.
+ * The user is responsible for freeing the memory by calling @ref nrf_cloud_fota_job_free.
+ */
 int nrf_cloud_rest_fota_execution_parse(const char *const response,
 					struct nrf_cloud_fota_job_info *const job);
 
+/** @brief Parses the PGPS response (REST and MQTT) from nRF Cloud */
 int nrf_cloud_parse_pgps_response(const char *const response,
 				  struct nrf_cloud_pgps_result *const result);
 
+/** @brief Adds common [network] modem info to the provided cJSON object */
 int nrf_cloud_json_add_modem_info(cJSON * const data_obj);
 
+/** @brief Builds a cellular positioning request string using the provided cell info.
+ * If successful, memory will be allocated for the output string and the user is
+ * responsible for freeing it using @ref cJSON_free.
+ */
 int nrf_cloud_format_cell_pos_req(struct lte_lc_cells_info const *const inf,
 				  size_t inf_cnt, char **string_out);
 
+/** @brief Builds a cellular positioning request in the provided cJSON object
+ * using the provided cell info
+ */
 int nrf_cloud_format_cell_pos_req_json(struct lte_lc_cells_info const *const inf,
 				       size_t inf_cnt, cJSON * const req_obj_out);
 
+/** @brief Builds a [single-cell] cellular positioning request in the provided cJSON object.
+ * Function obtains the necessary network info from the modem.
+ */
 int nrf_cloud_format_single_cell_pos_req_json(cJSON * const req_obj_out);
 
+/** @brief Parses the cellular positioning response (REST and MQTT) from nRF Cloud. */
 int nrf_cloud_parse_cell_pos_response(const char *const buf,
 				      struct nrf_cloud_cell_pos_result *result);
 
+/** @brief Obtains a pointer to the string at the specified index in the cJSON array.
+ * No memory is allocated, pointer is valid as long as the cJSON array is valid.
+ */
 int get_string_from_array(const cJSON * const array, const int index,
 			  char **string_out);
 
+/** @brief Obtains a pointer to the string of the specified key in the cJSON object.
+ * No memory is allocated, pointer is valid as long as the cJSON object is valid.
+ */
 int get_string_from_obj(const cJSON * const obj, const char *const key,
 			char **string_out);
+
+/** @brief Sends the cJSON object to nRF Cloud on the d2c topic */
+int json_send_to_cloud(cJSON * const request);
+
+/** @brief Creates a cJSON object containing the specified appId and messageType.
+ * If successful, user is responsible for calling @ref cJSON_Delete to free
+ * the cJSON object's memory.
+ */
+cJSON *json_create_req_obj(const char *const app_id, const char *const msg_type);
 
 #ifdef __cplusplus
 }
